Create directory Plugin/builder and move associated code inside it.
