+++ Plugin/builder/builder_NMake.cpp
