+++ Plugin/buildmanager.h
@@ -25,7 +25,7 @@
 #ifndef BUILDMANAGER_H
 #define BUILDMANAGER_H
 
-#include "builder.h"
+#include "builder/builder.h"
 #include "codelite_exports.h"
 #include "singleton.h"
 #include "wx/string.h"
